move InsertColumnsToken to pojo
